backend/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
numpy==1.26.4
pydantic==2.8.2
torch==2.6.0
python-dateutil==2.9.0.post0
